// api/src/llm/enums/gemini-text-models.enum.ts
export enum GeminiTextModel {
  // === Gemini 2.5 ===
  GEMINI_2_5_PRO = 'models/gemini-2.5-pro',
  GEMINI_2_5_FLASH = 'models/gemini-2.5-flash',

  // === Gemini 2.0 ===
  GEMINI_2_0_FLASH = 'models/gemini-2.0-flash',
  GEMINI_2_0_FLASH_LITE = 'models/gemini-2.0-flash-lite',
  GEMINI_2_0_FLASH_PREVIEW_IMAGE_GENERATION = 'models/gemini-2.0-flash-preview-image-generation',
  GEMINI_2_0_FLASH_LIVE = 'models/gemini-2.0-flash-live-001',

  // === Gemini 1.5 ===
  GEMINI_1_5_PRO = 'models/gemini-1.5-pro',
  GEMINI_1_5_FLASH = 'models/gemini-1.5-flash',
  GEMINI_1_5_FLASH_8B = 'models/gemini-1.5-flash-8b',
}